// frontend/src/api/generated/shop/schemas/shippingMethod.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * ShopDjango API
 * API for ShopDjango project
 * OpenAPI spec version: 1.0.0
 */

/**
 * Serializer for ShippingMethod model.
 */
export interface ShippingMethod {
  readonly id: number;
  /**
   * Name of the shipping method (e.g., 'Standard', 'Express')
   * @maxLength 100
   */
  name: string;
  /**
   * Shipping cost
   * @pattern ^-?\d{0,8}(?:\.\d{0,2})?$
   */
  price: string;
  /** Courier company providing this shipping method */
  courier: number;
}